refactor: move wallet interfaces to wallet/interfaces package
Per reviewer suggestion — extract all public interfaces (Wallet,
WalletERC20, WalletStableCoin and all StableCoin* sub-interfaces)
into a dedicated wallet/interfaces package. The wallet package
re-exports them as type aliases to maintain full backward compatibility:
existing code using wallet.Wallet, wallet.WalletStableCoin, etc. is
unaffected.
